firebase/functions/requirements.txt
flask==3.0.3
flask-cors==3.0.10
langchain==0.2.14
langchain_openai==0.1.22
faiss-cpu==1.8.0.post1
pickle-mixin==1.0.2
gunicorn==23.0.0
langchain_community==0.2.12
langchain_core==0.2.33
openai==1.41.0
langchain_text_splitters==0.2.2
git+https://github.com/m-bain/whisperx.git
rank_bm25
python-dotenv
Werkzeug==3.0.1
Flask-JWT-Extended==4.6.0
waitress==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
python-dotenv==0.19.0
firebase-admin
functions-framework
firebase-functions==0.1.0
